[BE] fix: library track
하드코딩된 user id를
req에서 뽑아 쓰도록 변경

+++ backend/controllers/library/tracks.ts
@@ -12,29 +12,45 @@ const getLibTracks = async (req: Request, res: Response): Promise<void> => {
     const result = await getUserLikeTracks(req.user.id);
     res.status(200).json({ Tracks: result });
   } catch (err) {
-    res.status(500).json({ statusCode: 500, message: err.message });
+    if (err === "Unauthorized") {
+      res.status(401).json({ message: "Unauthroized" });
+    } else {
+      res.status(500).json({ statusCode: 500, message: err.message });
+    }
   }
 };
 
 const postLibTracks = async (req: Request, res: Response): Promise<void> => {
-  const userId = 1; // decode jwt
-  const { id: trackId } = req.params;
   try {
+    if (!req.user) throw new Error("Unauthorized");
+    const userId = req.user.id;
+    const { id: trackId } = req.params;
+
     const msg = await postUserLikeTracks(userId, +trackId);
     res.status(200).json({ message: msg });
   } catch (err) {
-    res.status(500).json({ statusCode: 500, message: err.message });
+    if (err === "Unauthorized") {
+      res.status(401).json({ message: "Unauthroized" });
+    } else {
+      res.status(500).json({ statusCode: 500, message: err.message });
+    }
   }
 };
 
 const deleteLibTracks = async (req: Request, res: Response): Promise<void> => {
-  const userId = 1; // decode jwt
-  const { id: trackId } = req.params;
   try {
+    if (!req.user) throw new Error("Unauthorized");
+    const userId = req.user.id;
+    const { id: trackId } = req.params;
+
     const msg = await deleteUserLikeTracks(userId, +trackId);
     res.status(200).json({ message: msg });
   } catch (err) {
-    res.status(500).json({ statusCode: 500, message: err.message });
+    if (err === "Unauthorized") {
+      res.status(401).json({ message: "Unauthroized" });
+    } else {
+      res.status(500).json({ statusCode: 500, message: err.message });
+    }
   }
 };
 export { getLibTracks, postLibTracks, deleteLibTracks };
